fixing http://webcvs.ekg2.org/ekg2/plugins/logs/main.c.diff?r1=1.49&r2=1.50 argh.. 6 days broken logfiles.. whatever.
leafnode look in ruid we have receipient uid da.
so when we sent messages,
we should log our uid, not receipient uid, da?
because now, i have in my logfiles....
#v+
<smbuid> text that HE sent to us..
<smbuid> text that WE sent to him...
#v-
i think it's incorrect, because we ought have
#v+
<smbuid> text that HE sent to us..
<ouruid> text that WE sent to him...
#v+
da?
not smbuid twice, it mind if we send messages or we recved messages!!