ok, so remove valid_uid() it's never used.
change get_uid() get_nickname() to check only plugin which handle it.
however it may cause breakdown of ekg2.

So, i check some functions which calls it, and eventually after it do:

if (!return_value_of_getuid) return_value_of_getuid = value_passed_to_get_uid;
I really don't know if it should be this way.

Next, before this function we can check by valid_plugin_uid() if this is good uid for session at whole.
if not, we can display nicer message than: `user not found`

userlist not fixed yet. but today it should be ;)

many XXX's if someone think this should be fixed in 1way, message me. if in 2way, also.
I'll also think about it.. but in a while, ok?