Implementing session-locking, to avoid connecting to the same session more than one copy of EKG2 simultaneously. With special dedication to Greyer.
It is controlled by var session_locks. Allows two types of locking. Simpler
and more portable (2) just creates file on locking, and unlinks it on unlock.
If file exists before locking, locking fails and connect should be aborted.

The default one (1) uses additionally flock() on created file. The main
advantage of this is that if EKG2 dies, lock will die too, and we won't have
to remove it ,,by hand'' (i.e. using /session --unlock).

XXX: Do we need to check if flock() is available?