BIG FAT WARNING - LOOP MODIFICATION!
(I was too lazy to do this on #ifdef's)

Moved watch timeout checking after select() (i.e. inside select-checking
loop, if no data for watch is available). Why? Because we could remove
watch when it was already ready for read. If there was big lag between
creating watch and checking timeouts (logsqlite with very big db, nice 10,
and compilation in bg), it was possible that watch was removed without
even one check of its' state - i.e. 2 secs after creating watch it was
ready, but function hasn't returned for 60 secs, and watch was removed
because of timeout.