Speedup getting format.
Introduce gim_hash() which is better hash-function acording to ekg_hash_benchmark.c
[However it's a little slower...]

Change themes list_t to array of list_t (currently: with 256 items)
Each hash of theme is & with 0xff, so we decrease items to search from: O(~ 1095 :>) [yep, when we load all plugins ;>]
to max O(~ 12)

If we leave old hashing function - ekg_hash(), than we'll have about 133 items of array empty, (so it's bad) [vs 3 in gim_hash()]
(and avg O(~30))

However both hash functions are good, collisions with the same hash is rare. (In gim_hash() for now never)
so we just limit comparing integers, and going across list.

[Test results from ekg_hash_benchmark.c]

If you have better and faster hash function you can send it to ekg2-devel.

CUL.