Revert now-useless mkinstalldirs path setting change.
Revert darkjames' change from Tue Aug 22 22:30:54 2006 +0000 that was to fix
http://bugs.ekg2.org/task/49. The bug report is not accessible any more, but it
probably was about a wrong path to mkinstalldirs in the snapshots (possibly
caused by some incompatibility between automake and gettextize).

These days the snapshots are generated on a relatively recent system, with
gettext that comes with autopoint. Since we run autopoint --force, this change
was being reverted in the snapshots for some time now anyway. This (apart from
making the change effectively useless) also meant that if you ran autogen.sh, a
source file was getting modified. This made it easy to mistakenly commit the
change (and did happen a few times as seen from the file's log).