Fold Makefile and generate.sh into a new script for generating and cleaning ekg2book.
The Makefile files in ekg2book directories are the only non-generated Makefiles
in the source tree, so it is difficult to omit them when running
./autoclean.sh.

Also, having a single script reduces code duplication.

This script should be used from now on, instead of the Makefiles and/or
generate.sh scripts within ekg2book directories, as they will be going away
soon.