Fix segault in bad show_status_notavail format usage.
Jabber plugin called format_string with that format without an apparently
required argument. This caused a va_arg() call in va_format_string() to run
past the end of the list and return an invalid pointer, which upon dereference
caused a segfault.

It is unclear how to ensure that the formats match their arg count in the
general case. The fact that the format can require a differing number of
arguments depending on $LANG makes matters even more convoluted.