Made loading plugins from relative dirs optional.
Added a configure option which lets one disable loading plugins from
directories relative to $CWD. This is a security improvement for cases such as
where ekg2 is run from an untrusted directory, or where a malicious user
controls a user whose name is "plugins".

Also added a comment explaining the reason for this behaviour.