Qualify uid in struct userlist as const.
It needs to be const, because there are places where we point this member at
const data.

Qualifying it as const requires in turn some casting at xfree() time, but I
reckon it's better to do it there, rather than be const-incorrect for the rest
of lifetime of the data.