Don't rely on handler to free fstring_t on ui-window-print.
Change the API on window_print() and ui-window-print. Now the fstring_t
passed to that function/query needs to be freed by caller. If any
handler wishes to preserve or modify this data, it needs to use
fstring_dup().

The changed API simplifies handling of that information and avoids
random memleaks/segvs.