Make the fatal signal handlers async-signal-safe again.
Currently ekg2 likes to deadlock when it segfaults in a malloc/free because the
debug functions try to execute malloc/free again, and they are not necessarily
reentrant.

Remove a lot of the fancy things that ekg2 does on such signals and keep the
bare minimum with the goal of being able to produce usable corefiles again and
restoring terminal state to sanity.

Also, no longer change the current directory to the config dir - users
generally expect programs to dump core in the directory they were started in.